增加原图下载

+++ src/command/cmdWar3IconPanel.ts
@@ -40,21 +40,26 @@ export async function war3IconPanel(context: vscode.ExtensionContext) {
 		await dirExists(rootPath + `/${config_res}/ReplaceableTextures/CommandButtons`);
 		await dirExists(rootPath + `/${config_res}/ReplaceableTextures/PassiveButtons`);
 		await dirExists(rootPath + `/${config_res}/ReplaceableTextures/CommandButtonsDisabled`);
-		let btn_name = filename.substring(0, filename.length - 4);
-		let btn_png_path = mergeImages(pngPath, framedata[0], rootPath + `/${config_res}/ReplaceableTextures/CommandButtons/BTN` + btn_name + ".png");
-		blp2Image(btn_png_path, btn_png_path.substring(0,btn_png_path.length-4) + ".blp", 'blp');
-		fs.unlink(btn_png_path, () => { });
-
-		let pasbtn_png_path = mergeImages(pngPath, framedata[2], rootPath + `/${config_res}/ReplaceableTextures/PassiveButtons/PASBTN` + btn_name + ".png");
-		blp2Image(pasbtn_png_path, pasbtn_png_path.substring(0,pasbtn_png_path.length-4) + ".blp", 'blp');
-		fs.unlink(pasbtn_png_path, () => { });
-
-		let disbtn_png_path = mergeImages(pngPath, framedata[1], rootPath + `/${config_res}/ReplaceableTextures/CommandButtonsDisabled/DISBTN` + btn_name + ".png");
-		blp2Image(disbtn_png_path, disbtn_png_path.substring(0,disbtn_png_path.length-4) + ".blp", 'blp');
-		fs.unlink(disbtn_png_path, () => { });
-
-		if (is_remove_file) {
-			fs.unlink(pngPath, () => { });
+		if (framedata[0]=="原图") {
+			let btn_name = filename.substring(0, filename.length - 4);
+			let btn_png_path = mergeImages(pngPath, framedata[0], rootPath + `/${config_res}/ReplaceableTextures/CommandButtons/BTN` + btn_name + ".png");
+		}else{
+			let btn_name = filename.substring(0, filename.length - 4);
+			let btn_png_path = mergeImages(pngPath, framedata[0], rootPath + `/${config_res}/ReplaceableTextures/CommandButtons/BTN` + btn_name + ".png");
+			blp2Image(btn_png_path, btn_png_path.substring(0,btn_png_path.length-4) + ".blp", 'blp');
+			fs.unlink(btn_png_path, () => { });
+	
+			let pasbtn_png_path = mergeImages(pngPath, framedata[2], rootPath + `/${config_res}/ReplaceableTextures/PassiveButtons/PASBTN` + btn_name + ".png");
+			blp2Image(pasbtn_png_path, pasbtn_png_path.substring(0,pasbtn_png_path.length-4) + ".blp", 'blp');
+			fs.unlink(pasbtn_png_path, () => { });
+	
+			let disbtn_png_path = mergeImages(pngPath, framedata[1], rootPath + `/${config_res}/ReplaceableTextures/CommandButtonsDisabled/DISBTN` + btn_name + ".png");
+			blp2Image(disbtn_png_path, disbtn_png_path.substring(0,disbtn_png_path.length-4) + ".blp", 'blp');
+			fs.unlink(disbtn_png_path, () => { });
+	
+			if (is_remove_file) {
+				fs.unlink(pngPath, () => { });
+			}
 		}
 	}
 
@@ -94,7 +99,7 @@ export async function war3IconPanel(context: vscode.ExtensionContext) {
 
 				panel.webview.postMessage({ type: 'updateitems', text: replaceText, last_page: response.data.data.last_page });
 			});
-
+				
 			if (config_list) {
 				let replaceText = "";
 				let replaceTextStyle = "";
